+++ lib/widgets/widgets.dart
@@ -13,6 +13,7 @@ export 'src/clide_collapser_card.dart';
 export 'src/clide_column_hat.dart';
 export 'src/clide_code_block.dart';
 export 'src/clide_divider.dart';
+export 'src/clide_file_image.dart';
 export 'src/clide_filter_box.dart';
 export 'src/clide_lightbox.dart';
 export 'src/clide_markdown.dart';